Refactor financial components to integrate KeuanganSaldo and enhance data handling
- Replace KodeAkunNeraca with KeuanganSaldo in various components to streamline financial data management.
- Introduce a new method `tutupBuku` in JurnalkeuanganClass to check if the financial period is closed before allowing deletions.
- Update views and models to reflect changes in relationships and improve user interaction with financial records.
- Enhance the Pelunasan index view to display additional payment details and ensure proper validation for journal entries.

+++ app/Livewire/Manajemenstok/Pengadaanbrgdagang/Tagihan/Index.php
@@ -3,9 +3,10 @@
 namespace App\Livewire\Manajemenstok\Pengadaanbrgdagang\Tagihan;
 
 use Livewire\Component;
-use App\Models\PengadaanTagihan;
 use Livewire\Attributes\Url;
 use Livewire\WithPagination;
+use App\Models\PengadaanTagihan;
+use App\Class\JurnalkeuanganClass;
 
 class Index extends Component
 {
@@ -24,6 +25,10 @@ class Index extends Component
 
     public function delete($id)
     {
+        if (JurnalkeuanganClass::tutupBuku($this->bulan . '-01')) {
+            session()->flash('error', 'Pembukuan periode ini sudah ditutup');
+            return;
+        }
         $data = PengadaanTagihan::find($id);
         $data->delete();
         session()->flash('success', 'Berhasil menghapus data');
